requirements.txt
selenium==4.39.0
pytest==9.0.2
pytest-html==4.1.1
webdriver-manager==4.0.2
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1